// models/enums/PrivacyLevel.ts
export enum PrivacyLevel {
  PUBLIC = 'PUBLIC',         // 所有人可見
  REGISTERED = 'REGISTERED', // 只有註冊用戶可見
  APPLIED = 'APPLIED',       // 只有已申請的用戶可見
  ACCEPTED = 'ACCEPTED',     // 只有已接受的用戶可見
  PRIVATE = 'PRIVATE'        // 僅自己可見
}